requirements.txt
azure-search-documents==11.5.1
azure-storage-blob==12.23.1
